fix: add default values in application-local.yaml.
